delete activist
