+++ model/event_test.go
@@ -1,49 +1,11 @@
 package model
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
-	"github.com/dxe/adb/config"
-	"github.com/jmoiron/sqlx"
 	"github.com/stretchr/testify/assert"
 )
-
-func newTestDB() *sqlx.DB {
-	db := NewDB(config.DBTestDataSource())
-	WipeDatabase(db)
-
-	return db
-}
-
-func TestAutocompleteActivistsHandler(t *testing.T) {
-	db := newTestDB()
-	defer db.Close()
-
-	_, err := GetOrCreateUser(db, "User One")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	_, err = GetOrCreateUser(db, "User Two")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	gotNames := GetAutocompleteNames(db)
-	wantNames := []string{"User One", "User Two"}
-	fmt.Println(gotNames, wantNames)
-	if len(gotNames) != len(wantNames) {
-		t.Fatalf("gotNames and wantNames must have the same length.")
-	}
-
-	for i := range wantNames {
-		if gotNames[i] != wantNames[i] {
-			t.Fatalf("gotNames and wantNames must be equal")
-		}
-	}
-}
 
 func TestGetEvents(t *testing.T) {
 	db := newTestDB()
@@ -289,148 +251,4 @@ func TestDeleteEvents(t *testing.T) {
 		db.Select(&attendees, "select activist_id from event_attendance where event_id = 1"))
 
 	assert.Len(t, attendees, 0)
-}
-
-func TestGetUserEventData(t *testing.T) {
-	db := newTestDB()
-	defer db.Close()
-
-	u1, err := GetOrCreateUser(db, "Test User")
-	assert.NoError(t, err)
-
-	d1, err := time.Parse("2006-01-02", "2017-04-15")
-	assert.NoError(t, err)
-	d2, err := time.Parse("2006-01-02", "2017-04-16")
-	assert.NoError(t, err)
-	d3, err := time.Parse("2006-01-02", "2017-04-17")
-	assert.NoError(t, err)
-
-	// These events are intentionally out of order
-	insertEvents := []Event{{
-		ID:             1,
-		EventName:      "event one",
-		EventDate:      d2,
-		EventType:      "Working Group",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             2,
-		EventName:      "event two",
-		EventDate:      d1,
-		EventType:      "Working Group",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             3,
-		EventName:      "event three",
-		EventDate:      d3,
-		EventType:      "Working Group",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             4,
-		EventName:      "event four",
-		EventDate:      d3,
-		EventType:      "Working Group",
-		AddedAttendees: []User{u1},
-	}}
-	mustInsertAllEvents(t, db, insertEvents)
-
-	d, err := u1.GetUserEventData(db)
-	assert.NoError(t, err)
-
-	d.FirstEvent.Equal(d1)
-	d.LastEvent.Equal(d3)
-	assert.Equal(t, d.TotalEvents, 4)
-}
-
-func TestGetUserEventData_noEvents(t *testing.T) {
-	db := newTestDB()
-	defer db.Close()
-
-	u1, err := GetOrCreateUser(db, "Test User")
-	assert.NoError(t, err)
-
-	d, err := u1.GetUserEventData(db)
-	assert.NoError(t, err)
-
-	assert.Equal(t, d, UserEventData{
-		FirstEvent:  nil,
-		LastEvent:   nil,
-		TotalEvents: 0,
-	})
-}
-
-func mustInsertAllEvents(t *testing.T, db *sqlx.DB, events []Event) {
-	for _, e := range events {
-		_, err := InsertUpdateEvent(db, Event{
-			EventName:      e.EventName,
-			EventDate:      e.EventDate,
-			EventType:      e.EventType,
-			AddedAttendees: e.AddedAttendees,
-		})
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-}
-
-// TODO: Finish this test.
-func TestGetLeaderboardUsers(t *testing.T) {
-	db := newTestDB()
-	defer db.Close()
-
-	// Set up two events
-	u1, err := GetOrCreateUser(db, "Hello")
-	assert.NoError(t, err)
-
-	d2, err := time.Parse("2006-01-02", "2017-04-16")
-	assert.NoError(t, err)
-	var wantEvents = []Event{{
-		ID:             1,
-		EventName:      "event one",
-		EventDate:      d2,
-		EventType:      "Working Group",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             2,
-		EventName:      "event two",
-		EventDate:      d2,
-		EventType:      "Protest",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             3,
-		EventName:      "event three",
-		EventDate:      d2,
-		EventType:      "Key Event",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             4,
-		EventName:      "event four",
-		EventDate:      d2,
-		EventType:      "Community",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             5,
-		EventName:      "event five",
-		EventDate:      d2,
-		EventType:      "Outreach",
-		AddedAttendees: []User{u1},
-	}, {
-		ID:             6,
-		EventName:      "event six",
-		EventDate:      d2,
-		EventType:      "Sanctuary",
-		AddedAttendees: []User{u1},
-	}}
-
-	for _, e := range wantEvents {
-		_, err := InsertUpdateEvent(db, Event{
-			EventName:      e.EventName,
-			EventDate:      e.EventDate,
-			EventType:      e.EventType,
-			AddedAttendees: e.AddedAttendees,
-		})
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-
 }
