+++ templates/event_list.html
@@ -1,4 +1,6 @@
 {{template "header.html" .}}
+<!-- bootstrap js needed for toggle-dropdown -->
+<script src="/static/external/bootstrap/js/bootstrap.min.js"></script>
 <script src="/dist/event_list.js"></script>
 
 <div class="body-wrapper-wide event-list-content">
