+++ templates/header.html
@@ -13,7 +13,6 @@
     <link rel="stylesheet" type="text/css" href="/static/css/style.css?5">
     <link rel="stylesheet" type="text/css" href="/static/css/flash_message.css">
     <script src="/static/external/jquery-3.2.1.js"></script>
-    <script src="/static/external/bootstrap/js/bootstrap.min.js"></script>
 
     <script src="/dist/flash_message.js"></script>
   </head>
